Separate out common types into the types package
The forthcoming client and server will communicate via some public types.
This commit creates a new home for these public types; the types package.
These types already existed but were a little convoluted  as their
usage was overloaded somewhat.  Some of the methods they defined had
uses that were specific to one component.  For example, the ccvm package
doesn't need to know anything about parsing flags and the cmd package
doesn't need to know about marshalling and unmarshalling VMspecs.  For
this reason, some of the methods of the relocated types have been separated
out and moved into the packages where they logically belong.
